Track2/Track2/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.8.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.1
opencv-python-headless>=4.10.0.84
pillow>=10.4.0
numpy>=1.26.4
python-multipart>=0.0.9
requests>=2.32.0
reportlab>=4.2.2
pypdf>=5.1.0
python-docx>=1.1.2
python-pptx>=1.0.2
pytesseract>=0.3.10
pytest>=8.2.0
